feat(*): add delete vp node interface
1. Add delete vp node interface;
2. Place the node related interfaces in node.go.
