+++ go.mod
@@ -7,8 +7,8 @@ require (
 	github.com/docker/docker v1.13.1
 	github.com/golang/mock v1.4.3
 	github.com/meshplus/bitxhub-kit v1.1.2-0.20201023073721-052e6b89ea39
-	github.com/meshplus/bitxhub-model v1.1.2-0.20201023091417-b6445e44d535
+	github.com/meshplus/bitxhub-model v1.1.2-0.20201208133437-73fa50d5efed
 	github.com/sirupsen/logrus v1.5.0
 	github.com/stretchr/testify v1.4.0
-	google.golang.org/grpc v1.27.1
+	google.golang.org/grpc v1.33.1
 )
